Changes requested from PR
