+++ tests/api/api/feature_store/test_feature_sets.py
@@ -3,7 +3,8 @@ from uuid import uuid4
 
 from fastapi.testclient import TestClient
 from sqlalchemy.orm import Session
-from deepdiff import DeepDiff
+
+from .base import _patch_object, _assert_list_objects
 
 
 def _generate_feature_set(name):
@@ -50,61 +51,11 @@ def _generate_feature_set(name):
     }
 
 
-def _generate_feature_vector(name):
-    return {
-        "kind": "FeatureVector",
-        "metadata": {
-            "name": name,
-            "labels": {"owner": "nobody", "group": "dev"},
-            "tag": "latest",
-            "extra_metadata": 100,
-        },
-        "spec": {
-            "features": ["feature_set:*", "feature_set:something", "just_a_feature"],
-            "description": "just a bunch of features",
-            "extra_spec": True,
-        },
-        "status": {
-            "state": "created",
-            "target": "parquet",
-            "extra_status": {"field1": "value1", "field2": "value2"},
-        },
-    }
-
-
-def _assert_list_objects(
-    client: TestClient, entity_name, project, query, expected_number_of_entities
-):
-    entity_url_name = entity_name.replace("_", "-")
-    url = f"/api/projects/{project}/{entity_url_name}"
-    if query:
-        url = url + f"?{query}"
-    response = client.get(url)
-    assert response.status_code == HTTPStatus.OK.value
-    response_body = response.json()
-    assert entity_name in response_body
-    assert (
-        len(response_body[entity_name]) == expected_number_of_entities
-    ), f"wrong number of {entity_name} entities in response"
-    return response_body
-
-
 def _feature_set_create_and_assert(
     client: TestClient, project, feature_set, versioned=True
 ):
     response = client.post(
         f"/api/projects/{project}/feature-sets?versioned={versioned}", json=feature_set
-    )
-    assert response.status_code == HTTPStatus.OK.value
-    return response.json()
-
-
-def _feature_vector_create_and_assert(
-    client: TestClient, project, feature_vector, versioned=True
-):
-    response = client.post(
-        f"/api/projects/{project}/feature-vectors?versioned={versioned}",
-        json=feature_vector,
     )
     assert response.status_code == HTTPStatus.OK.value
     return response.json()
@@ -199,30 +150,6 @@ def test_feature_set_create_and_list(db: Session, client: TestClient) -> None:
     _assert_list_objects(client, "feature_sets", project_name, "label=color", 2)
     # handling multiple label queries has issues right now - needs to fix and re-run this test.
     # _assert_list_objects(client, "feature_sets", project_name, "label=owner=bob&label=color=red", 2)
-
-
-def _patch_object(
-    client: TestClient,
-    project_name,
-    name,
-    object_update,
-    object_url_path,
-    additive=False,
-):
-    patch_mode = "replace"
-    if additive:
-        patch_mode = "additive"
-    headers = {"x-mlrun-patch-mode": patch_mode}
-    response = client.patch(
-        f"/api/projects/{project_name}/{object_url_path}/{name}/references/latest",
-        json=object_update,
-        headers=headers,
-    )
-    assert response.status_code == HTTPStatus.OK.value
-    response = client.get(
-        f"/api/projects/{project_name}/{object_url_path}/{name}/references/latest"
-    )
-    return response.json()
 
 
 def test_feature_set_patch(db: Session, client: TestClient) -> None:
@@ -532,207 +459,3 @@ def test_features_list(db: Session, client: TestClient) -> None:
     assert (
         features_response["features"][1]["feature_set_digest"]["metadata"]["tag"] == tag
     )
-
-
-# There will be fields added (uid for example), but we don't allow any other changes
-def _assert_diff_contains_only_additions(original_object, new_object):
-    diff = DeepDiff(original_object, new_object, ignore_order=True,)
-    diff.pop("dictionary_item_added")
-    assert diff == {}
-
-
-def test_feature_vector_create(db: Session, client: TestClient) -> None:
-    project_name = f"prj-{uuid4().hex}"
-    name = "feature_set1"
-    feature_vector = _generate_feature_vector(name)
-    feature_vector["metadata"]["project"] = project_name
-
-    feature_vector_response = _feature_vector_create_and_assert(
-        client, project_name, feature_vector, True
-    )
-    _assert_diff_contains_only_additions(feature_vector, feature_vector_response)
-    uid = feature_vector_response["metadata"]["uid"]
-
-    feature_vector_response = client.get(
-        f"/api/projects/{project_name}/feature-vectors/{name}/references/latest"
-    )
-    assert feature_vector_response.status_code == HTTPStatus.OK.value
-    _assert_diff_contains_only_additions(feature_vector, feature_vector_response.json())
-
-    feature_vector_response = client.get(
-        f"/api/projects/{project_name}/feature-vectors/{name}/references/{uid}"
-    )
-    assert feature_vector_response.status_code == HTTPStatus.OK.value
-    # When querying by uid, tag will not be returned
-    feature_vector["metadata"].pop("tag")
-    _assert_diff_contains_only_additions(feature_vector, feature_vector_response.json())
-
-
-def test_list_feature_vectors(db: Session, client: TestClient) -> None:
-    project_name = f"prj-{uuid4().hex}"
-    count = 10
-    dead_count = 0
-    blue_lables_count = 0
-    ooga_name_count = 0
-    not_latest_count = 0
-    for i in range(count):
-        name = f"feature_vector_{i}"
-        feature_set = _generate_feature_vector(name)
-        # generate some variations
-        if i % 2 == 0:
-            feature_set["status"]["state"] = "dead"
-            dead_count = dead_count + 1
-        if i % 3 == 0:
-            feature_set["metadata"]["labels"] = {"owner": "somebody", "color": "blue"}
-            blue_lables_count = blue_lables_count + 1
-        if i % 4 == 0:
-            feature_set["metadata"]["name"] = f"ooga_booga_{i}"
-            ooga_name_count = ooga_name_count + 1
-        if i > 4:
-            feature_set["metadata"]["tag"] = "just_a_tag"
-            not_latest_count = not_latest_count + 1
-        _feature_vector_create_and_assert(client, project_name, feature_set)
-
-    _assert_list_objects(client, "feature_vectors", project_name, None, count)
-    _assert_list_objects(
-        client, "feature_vectors", project_name, "name=ooga", ooga_name_count
-    )
-    _assert_list_objects(
-        client, "feature_vectors", project_name, "label=color=blue", blue_lables_count
-    )
-    _assert_list_objects(client, "feature_vectors", project_name, "label=owner", count)
-    _assert_list_objects(
-        client, "feature_vectors", project_name, "state=dead", dead_count
-    )
-    _assert_list_objects(
-        client, "feature_vectors", project_name, "tag=just_a_tag", not_latest_count
-    )
-    _assert_list_objects(
-        client,
-        "feature_vectors",
-        project_name,
-        "state=dead&name=booga",
-        ooga_name_count,
-    )
-    _assert_list_objects(client, "feature_vectors", "wrong_project", None, 0)
-
-
-def _assert_store_feature_vector(
-    client: TestClient, project, name, reference, feature_vector, versioned=True
-):
-    response = client.put(
-        f"/api/projects/{project}/feature-vectors/{name}/references/{reference}?versioned={versioned}",
-        json=feature_vector,
-    )
-    assert response
-    return response.json()
-
-
-def test_feature_vector_store(db: Session, client: TestClient) -> None:
-    project_name = f"prj-{uuid4().hex}"
-    name = "feature_vector1"
-    feature_vector = _generate_feature_vector(name)
-
-    # Put a new object - verify it's created
-    response = _assert_store_feature_vector(
-        client, project_name, name, "latest", feature_vector
-    )
-    uid = response["metadata"]["uid"]
-    # Change fields that will not affect the uid, verify object is overwritten
-    feature_vector["status"]["state"] = "modified"
-
-    response = _assert_store_feature_vector(
-        client, project_name, name, "latest", feature_vector
-    )
-    assert response["metadata"]["uid"] == uid
-    assert response["status"]["state"] == "modified"
-
-    _assert_list_objects(client, "feature_vectors", project_name, f"name={name}", 1)
-
-    # Now modify in a way that will affect uid, add a field to the metadata.
-    # Since referencing the object as "latest", a new version (with new uid) should be created.
-    feature_vector["metadata"]["new_metadata"] = True
-    response = _assert_store_feature_vector(
-        client, project_name, name, "latest", feature_vector
-    )
-    modified_uid = response["metadata"]["uid"]
-    assert modified_uid != uid
-    assert response["metadata"]["new_metadata"] is True
-
-    _assert_list_objects(client, "feature_vectors", project_name, f"name={name}", 2)
-
-    # Do the same, but reference the object by its uid - this should fail the request
-    feature_vector["metadata"]["new_metadata"] = "something else"
-    response = client.put(
-        f"/api/projects/{project_name}/feature-vectors/{name}/references/{modified_uid}",
-        json=feature_vector,
-    )
-    assert response.status_code == HTTPStatus.BAD_REQUEST.value
-
-
-def test_feature_vector_patch(db: Session, client: TestClient) -> None:
-    project_name = f"prj-{uuid4().hex}"
-
-    name = "feature_vector_1"
-    feature_vector = _generate_feature_vector(name)
-    _feature_vector_create_and_assert(client, project_name, feature_vector)
-
-    # Update a feature-set
-    feature_vector_patch = {
-        "spec": {"extra_spec": "extra"},
-        "metadata": {"labels": {"new-label": "new-value", "owner": "someone-else"}},
-    }
-
-    patched_feature_vector = _patch_object(
-        client, project_name, name, feature_vector_patch, "feature-vectors"
-    )
-    patched_feature_vector_metadata = patched_feature_vector["metadata"]
-    assert (
-        # New label should be added
-        len(patched_feature_vector_metadata["labels"]) == 3
-        and "new-label" in patched_feature_vector_metadata["labels"]
-        and patched_feature_vector_metadata["labels"]["owner"] == "someone-else"
-    ), "update corrupted data - got wrong results for labels from DB after update"
-    assert (
-        patched_feature_vector["spec"]["extra_spec"]
-        == feature_vector_patch["spec"]["extra_spec"]
-    )
-
-    # use additive strategy, the new label should be added
-    feature_vector_patch = {
-        "metadata": {"labels": {"another": "one"}},
-    }
-    patched_feature_vector = _patch_object(
-        client,
-        project_name,
-        name,
-        feature_vector_patch,
-        "feature-vectors",
-        additive=True,
-    )
-    assert len(patched_feature_vector["metadata"]["labels"]) == 4
-
-
-def test_feature_vector_delete(db: Session, client: TestClient) -> None:
-    project_name = f"prj-{uuid4().hex}"
-    count = 5
-    for i in range(count):
-        name = f"feature_vector_{i}"
-        feature_vector = _generate_feature_vector(name)
-        _feature_vector_create_and_assert(client, project_name, feature_vector)
-
-    _assert_list_objects(client, "feature_vectors", project_name, None, count)
-
-    # Delete the last feature vector
-    response = client.delete(
-        f"/api/projects/{project_name}/feature-vectors/feature_vector_{count-1}"
-    )
-    assert response.status_code == HTTPStatus.NO_CONTENT.value
-    _assert_list_objects(client, "feature_vectors", project_name, None, count - 1)
-
-    # Delete the first fs
-    response = client.delete(
-        f"/api/projects/{project_name}/feature-vectors/feature_vector_0"
-    )
-    assert response.status_code == HTTPStatus.NO_CONTENT.value
-    _assert_list_objects(client, "feature_vectors", project_name, None, count - 2)
